Replace arrays in ModelManager with HeaderModelDefinitions

+++ Labs/Lab_Project/Model/ModelFactory.h
@@ -1,7 +1,5 @@
 #pragma once
 
-#include <string>
-
 #include "Model.h"
 
 
@@ -10,11 +8,11 @@ class ModelFactory
 {
 public:
     // Creates a new model from a data array containing data in format: 3 position coordinates, 3 normal coordinates
-    static Model* createFrom3Pos3Norm(const float* data, unsigned long length);
+    static Model* createFrom3Pos3Norm(const float* data, size_t length);
 
     // Creates a new model from a data array containing data in format: 3 position coordinates, 3 normal coordinates and 2 texture coordinates
-    static Model* createFrom3Pos3Norm2Tex(const float* data, unsigned long length);
+    static Model* createFrom3Pos3Norm2Tex(const float* data, size_t length);
 
     // Creates a new model from a data array containing data in format: 3 position coordinates
-    static Model* createFrom3Pos(const float* data, unsigned long length);
+    static Model* createFrom3Pos(const float* data, size_t length);
 };
